Balancer/pyproject.toml
[tool.poetry]
name = "balancer"
version = "0.1.0"
description = ""
authors = ["''"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
loguru = "0.7.2"
nats-py = "2.9.0"
uvloop = "0.21.0"
orjson = "3.10.10"
python-decouple = "3.8"
setuptools = "75.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
